% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svGUI-package.R
\docType{package}
\name{svGUI-package}
\alias{svGUI-package}
\title{SciViews - Functions to Manage GUIs in R}
\description{
The SciViews svGUI package provides function to implement GUI
(Graphical User Interface) in R. It is independent from any particular GUI
toolkit, centralize info about GUI elements currently used, and dispatch GUI
functions to the particular toolkits used. It use the S3 object \code{gui}.
}
\section{Methods}{


\code{gui} implements two methods:

\describe{
\item{\code{print}}{Give information about the current state of the GUI}
\item{\code{$}}{Give access to various GUI properties.}
}
}

\section{Important functions}{

\code{\link[=gui_add]{gui_add()}} and \code{\link[=gui_change]{gui_change()}} for construction and management,
\code{\link[=dont_ask]{dont_ask()}} to determine if the GUI can interrupt R to ask something to the
user.
}

