% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\alias{$.gui}
\alias{print.gui}
\alias{is.gui}
\title{A GUI object.}
\usage{
\method{$}{gui}(gui, x)

\method{print}{gui}(x, ...)

is.gui(x)
}
\arguments{
\item{gui}{A \code{gui} object..}

\item{x}{A function for \code{$}, or an object for the others.}

\item{...}{Further arguments (not used yet).}
}
\description{
The \code{gui} object contains and manages GUI-related data.
}
\examples{
# Create a GUI
gui_add("myGUI")
is.gui(myGUI)
myGUI
gui_remove("myGUI")
}
\seealso{
\code{\link[=gui_add]{gui_add()}}
}
\concept{
GUI API implementation
}
\keyword{misc}
