\name{svGUI-package}
\alias{svGUI-package}
\alias{svGUI}
\docType{package}

\title{ SciViews GUI API - Functions to manage GUI client }
\description{
  The SciViews svGUI package provides function to communicate with a SciViews
  GUI client. Currently, it obnly supports SciViews Komodo, which is Komodo Edit
  or IDE with the SciViews-K plugin (see http://www.sciviews.org/SciViews-K).
}
\details{
  \tabular{ll}{
    Package: \tab svGUI\cr
    Type: \tab Package\cr
    Version: \tab 0.9-50\cr
    Date: \tab 2010-10-01\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API communication with client Komodo }
