\name{svGUI-package}
\alias{svGUI-package}
\alias{svGUI}
\docType{package}

\title{ SciViews GUI API - Functions to manage GUIs in R }
\description{
  The SciViews svGUI package provides function to implement GUI (Graphical User
  Interface) in R. It is independent from any particular GUI toolkit, centralize
  info about GUI elements currently used, and dispatch GUI functions to the
  particular toolkits used.
}
\details{
  \tabular{ll}{
    Package: \tab svGUI\cr
    Type: \tab Package\cr
    Version: \tab 0.9-55\cr
    Date: \tab 2014-03-02\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API implementation of GUI elements }
